Add support for simulation components and demo component.
