+++ src/plugins/simulator/components/simulator-tile.scss
@@ -1,19 +1,29 @@
 @import "../../../components/vars.sass";
 
 .simulator-content-container {
-  width: 100%;
   height: 100%;
+  width: 100%;
   text-align: left;
   overflow: auto;
 
   .simulator-content {
+    height: 100%;
+    width: 100%;
+
     .simulator-variables {
       display: flex;
       font-family: 'Ubunutu', sans-serif;
-      height: 100%;
+      height: 20px;
       justify-content: space-between;
-      padding: 45px 20px;
-      position: absolute;
+      margin: 45px 20px 10px;
+      width: 100%;
+    }
+
+    .simulator-component-container {
+      align-content: center;
+      display: flex;
+      height: calc(100% - 75px);
+      justify-content: center;
       width: 100%;
     }
   }
