+++ src/plugins/simulator/components/simulator-variable.scss
@@ -1,7 +1,6 @@
 .simulator-variable {
   align-items: center;
   display: flex;
-  height: 20px;
   // For now, use a fixed width to prevent the variables from shifting as their values change
   width: 180px;
