+++ src/plugins/simulator/simulations/brainwaves-gripper.scss
@@ -1,0 +1,31 @@
+.bwg-component {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+
+  .lightbulb {
+    height: 40px;
+    width: 40px;
+
+    &.on {
+      background-color: yellow;
+    }
+    &.off {
+      background-color: black;
+    }
+  }
+
+  .emg-track {
+    margin-top: 20px;
+    position: relative;
+    width: 150px;
+
+    .emg {
+      background-color: blue;
+      height: 20px;
+      position: absolute;
+      width: 20px;
+    }
+  }
+}
