--- src/plugins/simulator/simulations/brainwaves-gripper.ts
@@ -1,27 +1,0 @@
-import { ISimulation } from "./simulation-types";
-
-export const kBrainwavesKey = "EMG_and_claw";
-const kEMGKey = "input_EMG";
-
-export const brainwavesGripperSimulation: ISimulation = {
-  delay: 200,
-  variables: [
-    {
-      name: kEMGKey,
-      value: 0
-    },
-    {
-      name: "output_LightBulb",
-      value: 0
-    }
-  ],
-  values: {
-    [kEMGKey]: [
-      0, 0.3090169943749474, 0.5877852522924731, 0.8090169943749475, 0.9510565162951535,
-      1, 0.9510565162951536, 0.8090169943749475, 0.5877852522924732, 0.3090169943749475,
-      1.2246467991473532e-16, -0.30901699437494773, -0.587785252292473, -0.8090169943749473,
-      -0.9510565162951535, -1, -0.9510565162951536, -0.8090169943749476, -0.5877852522924734,
-      -0.3090169943749477, -2.4492935982947064e-16
-    ]
-  }
-};
