--- src/plugins/simulator/simulations/simulation-types.ts
@@ -1,7 +1,0 @@
-import { VariableSnapshot } from "@concord-consortium/diagram-view";
-
-export interface ISimulation {
-  delay: number,
-  variables: VariableSnapshot[],
-  values: Record<string, number[]>
-}
